move location panel constants
