+++ ui/shared/components/panel/search/constants.js
@@ -2,20 +2,13 @@ import React from 'react'
 import { Form } from 'semantic-ui-react'
 import styled from 'styled-components'
 
-import { CreateLocusListButton } from 'shared/components/buttons/LocusListButtons'
 import { RadioGroup, AlignedBooleanCheckbox, Select, InlineToggle } from 'shared/components/form/Inputs'
-import { snakecaseToTitlecase, camelcaseToTitlecase } from 'shared/utils/stringUtils'
+import { snakecaseToTitlecase } from 'shared/utils/stringUtils'
 import {
-  LOCUS_LIST_ITEMS_FIELD,
   ORDERED_PREDICTOR_FIELDS,
-  PANEL_APP_CONFIDENCE_LEVELS,
   SPLICE_AI_FIELD,
   predictorColorRanges,
 } from 'shared/utils/constants'
-
-import LocusListItemsFilter from './LocusListItemsFilter'
-import PaMoiSelector from './PaMoiSelector'
-import PaLocusListSelector from './PaLocusListSelector'
 
 export const THIS_CALLSET_FREQUENCY = 'callset'
 export const SV_CALLSET_FREQUENCY = 'sv_callset'
@@ -74,74 +67,6 @@ export const SV_FREQUENCIES = [
 ]
 
 export const FREQUENCIES = [...SNP_FREQUENCIES, ...MITO_FREQUENCIES, ...SV_FREQUENCIES]
-
-export const LOCUS_FIELD_NAME = 'locus'
-export const PANEL_APP_FIELD_NAME = 'panelAppItems'
-export const SELECTED_MOIS_FIELD_NAME = 'selectedMOIs'
-const VARIANT_FIELD_NAME = 'rawVariantItems'
-const PANEL_APP_COLORS = [...new Set(
-  Object.entries(PANEL_APP_CONFIDENCE_LEVELS).sort((a, b) => b[0] - a[0]).map(config => config[1]),
-)]
-export const LOCATION_FIELDS = [
-  {
-    name: LOCUS_LIST_ITEMS_FIELD.name,
-    label: LOCUS_LIST_ITEMS_FIELD.label,
-    labelHelp: LOCUS_LIST_ITEMS_FIELD.labelHelp,
-    component: LocusListItemsFilter,
-    width: 9,
-    shouldShow: locus => !locus[PANEL_APP_FIELD_NAME],
-    shouldDisable: locus => !!locus[VARIANT_FIELD_NAME],
-  },
-  ...PANEL_APP_COLORS.map(color => ({
-    key: color,
-    name: `${PANEL_APP_FIELD_NAME}.${color}`,
-    iconColor: color,
-    label: color === 'none' ? 'Genes' : `${camelcaseToTitlecase(color)} Genes`,
-    labelHelp: 'A list of genes, can be separated by commas or whitespace',
-    component: LocusListItemsFilter,
-    filterComponent: PaLocusListSelector,
-    width: 3,
-    shouldShow: locus => !!locus[PANEL_APP_FIELD_NAME],
-    shouldDisable: locus => !!locus[VARIANT_FIELD_NAME],
-    color,
-  })),
-  {
-    name: VARIANT_FIELD_NAME,
-    label: 'Variants',
-    labelHelp: 'A list of variants. Can be separated by commas or whitespace. Variants can be represented by rsID or in the form <chrom>-<pos>-<ref>-<alt>',
-    component: LocusListItemsFilter,
-    width: 4,
-    shouldDisable: locus => !!locus[LOCUS_LIST_ITEMS_FIELD.name] || !!locus[PANEL_APP_FIELD_NAME],
-  },
-  {
-    name: SELECTED_MOIS_FIELD_NAME,
-    label: 'Modes of Inheritance',
-    labelHelp: 'Filter the Gene List based on Modes of Inheritance from Panel App',
-    component: LocusListItemsFilter,
-    filterComponent: PaMoiSelector,
-    width: 6,
-    shouldDisable: locus => !!locus[VARIANT_FIELD_NAME],
-    shouldShow: locus => !!locus[PANEL_APP_FIELD_NAME],
-  },
-  {
-    name: 'create',
-    fullFieldValue: true,
-    component: LocusListItemsFilter,
-    control: CreateLocusListButton,
-    width: 4,
-    shouldShow: locus => !locus[PANEL_APP_FIELD_NAME],
-    shouldDisable: locus => !locus[LOCUS_LIST_ITEMS_FIELD.name],
-  },
-  {
-    name: 'excludeLocations',
-    component: LocusListItemsFilter,
-    filterComponent: AlignedBooleanCheckbox,
-    label: 'Exclude locations',
-    labelHelp: 'Search for variants not in the specified genes/ intervals',
-    width: 10,
-    shouldDisable: locus => !!locus[VARIANT_FIELD_NAME],
-  },
-]
 
 const REQUIRE_SCORE_FIELD = {
   name: 'requireScore',
